chore(connect): split up onCallFirmwareUpdate
