+++ packages/connect/src/firmware/getBinary.ts
@@ -1,8 +1,11 @@
-import { VersionArray } from '@trezor/device-utils';
-import { removeTrailingSlashes } from '@trezor/utils';
+import { FirmwareType, VersionArray } from '@trezor/device-utils';
+import { Log, removeTrailingSlashes } from '@trezor/utils';
 
 import { parseFirmwareHeaders } from './parseFirmwareHeaders';
 import { ERRORS } from '../constants';
+import { FirmwareUpdateParams } from './types';
+import { getFirmwareLocation, getReleaseByVersion } from '../data/firmwareInfo';
+import type { Device } from '../device/Device';
 import { BinaryInfo } from '../types';
 import { httpRequest } from '../utils/assets';
 
@@ -39,10 +42,74 @@ export const getBinary = async ({
     };
 };
 
+// Used in checkFirmwareHash
 export const getBinaryOptional = async (params: GetBinaryParams) => {
     try {
         return await getBinary(params);
     } catch {
         return null;
     }
+};
+
+type BinaryHelperParams = {
+    device: Device;
+    params: FirmwareUpdateParams;
+    firmwareType: FirmwareType;
+    isIntermediary: boolean;
+    log: Log;
+};
+
+// Used in onCallFirmwareUpdate
+export const getBinaryHelper = async ({
+    device,
+    params,
+    firmwareType,
+    isIntermediary,
+    log,
+}: BinaryHelperParams): Promise<BinaryInfo> => {
+    if (params.binary) {
+        return Promise.resolve({
+            binary: params.binary,
+            binaryVersion: parseFirmwareHeaders(Buffer.from(params.binary)).version,
+            releaseVersion: undefined,
+        });
+    }
+
+    if (!device.firmwareReleaseConfigInfo) {
+        throw ERRORS.TypedError('Runtime', 'device.firmwareReleaseConfigInfo is not set');
+    }
+    const deviceModel = device.features?.internal_model;
+
+    const {
+        release: { version },
+        intermediary,
+    } = device.firmwareReleaseConfigInfo;
+    log.debug(
+        'onCallFirmwareUpdate loading binary',
+        'isIntermediary',
+        isIntermediary,
+        'version',
+        version,
+        'firmwareType',
+        firmwareType,
+        'deviceModel',
+        deviceModel,
+    );
+
+    // We want to get the path url to the release from the specific release we want, in `firmwareReleaseConfigInfo`
+    // we have only information about latest release of current FirmwareType but if we want to change from
+    // Universal to BitcoinOnly then using url from `firmwareReleaseConfigInfo` would not work.
+    const release = await getReleaseByVersion(device.features, version, firmwareType);
+    if (!release) {
+        throw new Error('Missing Firmware release for device');
+    }
+    const { baseUrl, path } = getFirmwareLocation({
+        firmwareVersion: version,
+        remotePath: release.url,
+        deviceModel,
+        firmwareType,
+        intermediaryVersion: isIntermediary && intermediary ? intermediary.version : undefined,
+    });
+
+    return getBinary({ baseUrl, path, version });
 };
