+++ scripts/circular-dependencies/madge-snapshot.txt
@@ -87,22 +87,22 @@ packages/connect/src/data/firmwareInfo.ts > packages/connect/src/data/DataManage
 packages/connect/src/data/firmwareInfo.ts > packages/connect/src/utils/firmwareUtils.ts
 packages/connect/src/device/Device.ts > packages/connect/src/device/DeviceCommands.ts > packages/connect/src/device/DeviceCurrentSession.ts
 packages/connect/src/device/Device.ts > packages/connect/src/device/StateStorage.ts
+packages/connect/src/device/Device.ts > packages/connect/src/device/checks/checkFirmwareHashWithRetries.ts
+packages/connect/src/device/Device.ts > packages/connect/src/device/checks/checkFirmwareHashWithRetries.ts > packages/connect/src/device/checks/checkFirmwareHash.ts
+packages/connect/src/device/Device.ts > packages/connect/src/device/checks/checkFirmwareHashWithRetries.ts > packages/connect/src/device/checks/checkFirmwareHash.ts > packages/connect/src/firmware/index.ts > packages/connect/src/firmware/getBinary.ts
+packages/connect/src/device/Device.ts > packages/connect/src/device/checks/checkFirmwareHashWithRetries.ts > packages/connect/src/device/checks/checkFirmwareHash.ts > packages/connect/src/firmware/index.ts > packages/connect/src/firmware/getBinary.ts > packages/connect/src/firmware/types.ts > packages/connect/src/types/index.ts > packages/connect/src/types/api/index.ts > packages/connect/src/types/api/off.ts > packages/connect/src/events/ui-request.ts > packages/connect/src/core/AbstractMethod.ts
 packages/connect/src/device/Device.ts > packages/connect/src/device/thp/index.ts
 packages/connect/src/device/Device.ts > packages/connect/src/device/thp/index.ts > packages/connect/src/device/thp/handshake.ts
 packages/connect/src/device/Device.ts > packages/connect/src/device/thp/index.ts > packages/connect/src/device/thp/handshake.ts > packages/connect/src/device/thp/thpCall.ts
 packages/connect/src/device/Device.ts > packages/connect/src/device/thp/index.ts > packages/connect/src/device/thp/pairing.ts
 packages/connect/src/device/Device.ts > packages/connect/src/device/thp/index.ts > packages/connect/src/device/thp/session.ts
-packages/connect/src/device/Device.ts > packages/connect/src/device/workflow/checkFirmwareHashWithRetries.ts
-packages/connect/src/device/Device.ts > packages/connect/src/device/workflow/checkFirmwareHashWithRetries.ts > packages/connect/src/device/workflow/checkFirmwareHash.ts
-packages/connect/src/device/Device.ts > packages/connect/src/device/workflow/checkFirmwareHashWithRetries.ts > packages/connect/src/device/workflow/checkFirmwareHash.ts > packages/connect/src/api/firmware/index.ts > packages/connect/src/api/firmware/getBinary.ts > packages/connect/src/types/index.ts > packages/connect/src/types/api/index.ts > packages/connect/src/types/api/off.ts > packages/connect/src/events/ui-request.ts > packages/connect/src/core/AbstractMethod.ts
-packages/connect/src/device/Device.ts > packages/connect/src/device/workflow/checkFirmwareHashWithRetries.ts > packages/connect/src/device/workflow/checkFirmwareHash.ts > packages/connect/src/api/firmware/index.ts > packages/connect/src/api/firmware/uploadFirmware.ts
 packages/connect/src/device/Device.ts > packages/connect/src/device/workflow/handshake.ts > packages/connect/src/types/workflow.ts
 packages/connect/src/events/index.ts > packages/connect/src/events/call.ts > packages/connect/src/device/Device.ts
 packages/connect/src/events/index.ts > packages/connect/src/events/call.ts > packages/connect/src/device/Device.ts > packages/connect/src/device/DeviceCommands.ts > packages/connect/src/device/DeviceCurrentSession.ts
+packages/connect/src/events/index.ts > packages/connect/src/events/call.ts > packages/connect/src/device/Device.ts > packages/connect/src/device/checks/checkFirmwareHashWithRetries.ts > packages/connect/src/device/checks/checkFirmwareHash.ts > packages/connect/src/firmware/index.ts > packages/connect/src/firmware/getBinary.ts > packages/connect/src/firmware/types.ts
+packages/connect/src/events/index.ts > packages/connect/src/events/call.ts > packages/connect/src/device/Device.ts > packages/connect/src/device/checks/checkFirmwareHashWithRetries.ts > packages/connect/src/device/checks/checkFirmwareHash.ts > packages/connect/src/firmware/index.ts > packages/connect/src/firmware/getBinary.ts > packages/connect/src/firmware/types.ts > packages/connect/src/types/index.ts > packages/connect/src/types/api/index.ts > packages/connect/src/types/api/off.ts > packages/connect/src/events/ui-request.ts > packages/connect/src/core/AbstractMethod.ts
 packages/connect/src/events/index.ts > packages/connect/src/events/call.ts > packages/connect/src/device/Device.ts > packages/connect/src/device/thp/index.ts > packages/connect/src/device/thp/handshake.ts > packages/connect/src/device/thp/thpCall.ts
 packages/connect/src/events/index.ts > packages/connect/src/events/call.ts > packages/connect/src/device/Device.ts > packages/connect/src/device/thp/index.ts > packages/connect/src/device/thp/pairing.ts
-packages/connect/src/events/index.ts > packages/connect/src/events/call.ts > packages/connect/src/device/Device.ts > packages/connect/src/device/workflow/checkFirmwareHashWithRetries.ts > packages/connect/src/device/workflow/checkFirmwareHash.ts > packages/connect/src/api/firmware/index.ts > packages/connect/src/api/firmware/getBinary.ts > packages/connect/src/types/index.ts > packages/connect/src/types/api/index.ts > packages/connect/src/types/api/off.ts > packages/connect/src/events/ui-request.ts > packages/connect/src/core/AbstractMethod.ts
-packages/connect/src/events/index.ts > packages/connect/src/events/call.ts > packages/connect/src/device/Device.ts > packages/connect/src/device/workflow/checkFirmwareHashWithRetries.ts > packages/connect/src/device/workflow/checkFirmwareHash.ts > packages/connect/src/api/firmware/index.ts > packages/connect/src/api/firmware/uploadFirmware.ts
 packages/ipc-proxy/src/proxy-handler.ts > packages/ipc-proxy/src/validateIpcMessage.ts
 packages/product-components/src/components/SelectAssetModal/SelectAssetModal.tsx > packages/product-components/src/components/SelectAssetModal/AssetItem.tsx
 packages/suite-desktop-core/e2e/support/analytics.ts > packages/suite-desktop-core/e2e/support/fixtures.ts
